+++ src/noahtheduke/splint/rules/style/redundant_nested_call.clj
@@ -9,13 +9,8 @@
 
 (set! *warn-on-reflection* true)
 
-(def relevant-call?
-  #{'* '*' '+ '+'
-    'and 'comp 'concat 'every-pred 'lazy-cat
-    'max 'merge 'min 'or 'some-fn 'str})
-
 (defrule style/redundant-nested-call
-  "Some clojure.core functions and macros take a variable number of args, so there's no need to nest calls.
+  "Some clojure.core functions and macros take a variable number of args, so there's no need to nest calls. To check non-clojure.core functions, they can be added to the config under key `:fn-names`: `style/redundant-nested-call {:fn-names [foo]}`.
 
   > [!NOTE]
   > This can have performance implications in certain hot-loops.
@@ -29,9 +24,18 @@
   ; prefer
   (+ 1 2 3 4)
   (comp :foo :bar :qux :ply)
+
+  ; with `:fn-names [foo]`
+  ; avoid
+  (foo 1 2 (foo 3 4))
+
+  ; prefer
+  (foo 1 2 3 4)
   "
-  {:pattern '((? call relevant-call?) ?+args (?call ?+others))
-   :on-match (fn [ctx rule form {:syms [?call ?args ?others]}]
-               (let [new-form (list* ?call (concat ?args ?others))]
-                 (->diagnostic ctx rule form {:message (format "Redundant nested call: `%s`." ?call)
-                                              :replace-form new-form})))})
+  {:pattern '(?fun ?+args (?fun ?+others))
+   :on-match (fn [ctx {{:keys [fn-names]} :config :as rule} form {:syms [?fun ?args ?others]}]
+               (when (and (symbol? ?fun)
+                       (contains? fn-names (symbol (name ?fun))))
+                 (let [new-form (list* ?fun (concat ?args ?others))]
+                   (->diagnostic ctx rule form {:message (format "Redundant nested call: `%s`." ?fun)
+                                                :replace-form new-form}))))})